[Grupo 14] Costumer Products - Frontend/Backend

+++ back-end/src/database/migrations/20220920203506-create-sales.js
@@ -9,7 +9,7 @@ module.exports = {
         type: Sequelize.INTEGER
       },
       totalPrice: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.DECIMAL(9,2),
         allowNull: false,
       },
       deliveryAddress: {
